Disable refetch for reactQueries (#11200)
