+++ airbyte-webapp/src/App.tsx
@@ -1,8 +1,6 @@
 import React, { Suspense } from "react";
 import { ThemeProvider } from "styled-components";
 import { IntlProvider } from "react-intl";
-import { CacheProvider } from "rest-hooks";
-import { QueryClient, QueryClientProvider } from "react-query";
 import { BrowserRouter as Router } from "react-router-dom";
 
 import en from "./locales/en.json";
@@ -26,6 +24,7 @@ import {
   windowConfigProvider,
 } from "./config";
 import { WorkspaceServiceProvider } from "./services/workspaces/WorkspacesService";
+import { StoreProvider } from "views/common/StoreProvider";
 
 const StyleProvider: React.FC = ({ children }) => (
   <ThemeProvider theme={theme}>
@@ -44,20 +43,6 @@ const I18NProvider: React.FC = ({ children }) => (
   >
     {children}
   </IntlProvider>
-);
-
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      suspense: true,
-    },
-  },
-});
-
-const StoreProvider: React.FC = ({ children }) => (
-  <CacheProvider>
-    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-  </CacheProvider>
 );
 
 const configProviders: ValueProvider<Config> = [
