+++ airbyte-webapp/src/packages/cloud/App.tsx
@@ -1,9 +1,7 @@
 import React, { Suspense } from "react";
 import { ThemeProvider } from "styled-components";
 import { IntlProvider } from "react-intl";
-import { CacheProvider } from "rest-hooks";
 import { BrowserRouter as Router } from "react-router-dom";
-import { QueryClient, QueryClientProvider } from "react-query";
 
 import en from "locales/en.json";
 import cloudLocales from "packages/cloud/locales/en.json";
@@ -20,6 +18,7 @@ import { AuthenticationProvider } from "packages/cloud/services/auth/AuthService
 import { AppServicesProvider } from "./services/AppServicesProvider";
 import { IntercomProvider } from "./services/thirdParty/intercom/IntercomProvider";
 import { ConfigProvider } from "./services/ConfigProvider";
+import { StoreProvider } from "views/common/StoreProvider";
 
 const messages = Object.assign({}, en, cloudLocales);
 
@@ -42,36 +41,20 @@ const StyleProvider: React.FC = ({ children }) => (
   </ThemeProvider>
 );
 
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      suspense: true,
-    },
-  },
-});
-
-const StoreProvider: React.FC = ({ children }) => (
-  <CacheProvider>
-    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-  </CacheProvider>
-);
-
 const Services: React.FC = ({ children }) => (
-  <ConfigProvider>
-    <AnalyticsProvider>
-      <ApiErrorBoundary>
-        <NotificationServiceProvider>
-          <FeatureService>
-            <AppServicesProvider>
-              <AuthenticationProvider>
-                <IntercomProvider>{children}</IntercomProvider>
-              </AuthenticationProvider>
-            </AppServicesProvider>
-          </FeatureService>
-        </NotificationServiceProvider>
-      </ApiErrorBoundary>
-    </AnalyticsProvider>
-  </ConfigProvider>
+  <AnalyticsProvider>
+    <ApiErrorBoundary>
+      <NotificationServiceProvider>
+        <FeatureService>
+          <AppServicesProvider>
+            <AuthenticationProvider>
+              <IntercomProvider>{children}</IntercomProvider>
+            </AuthenticationProvider>
+          </AppServicesProvider>
+        </FeatureService>
+      </NotificationServiceProvider>
+    </ApiErrorBoundary>
+  </AnalyticsProvider>
 );
 
 const App: React.FC = () => {
@@ -81,11 +64,13 @@ const App: React.FC = () => {
         <I18NProvider>
           <StoreProvider>
             <Suspense fallback={<LoadingPage />}>
-              <Router>
-                <Services>
-                  <Routing />
-                </Services>
-              </Router>
+              <ConfigProvider>
+                <Router>
+                  <Services>
+                    <Routing />
+                  </Services>
+                </Router>
+              </ConfigProvider>
             </Suspense>
           </StoreProvider>
         </I18NProvider>
